Automating passwordless ssh setup
Usage:
Populate the SSH_HOSTS in gf_perf_config with the IP addresses that need to
be part of the passwordless ssh trusted network. pwl_ssh tries two different
passwords, and fails if neither of them can provide access to the system.
Run automate_pwl_ssh. It attempts to install expect if not already found on
the system, failing which, it exits with a suitable message.